Fix for #1564. Improved custom aggregates functionality. (#1823)
* Fix for #1564. Improved custom aggregates functionality.

* Simplified argument check.
