+++ Source/LinqToDB/Expressions/InternalExtensions.cs
@@ -845,6 +845,12 @@ namespace LinqToDB.Expressions
 			return expr;
 		}
 
+		public static Expression SkipMethodChain(this Expression expr, MappingSchema mappingSchema)
+		{
+			var result = Sql.ExtensionAttribute.ExcludeExtensionChain(mappingSchema, expr);
+			return result;
+		}
+
 		public static Dictionary<Expression,Expression> GetExpressionAccessors(this Expression expression, Expression path)
 		{
 			var accessors = new Dictionary<Expression,Expression>();
@@ -894,6 +900,8 @@ namespace LinqToDB.Expressions
 		{
 			if (expr == null)
 				return null;
+
+			expr = expr.SkipMethodChain(mapping);
 
 			switch (expr.NodeType)
 			{
@@ -995,6 +1003,14 @@ namespace LinqToDB.Expressions
 			return false;
 		}
 
+		public static bool IsExtensionMethod(this MethodCallExpression methodCall, MappingSchema mapping)
+		{
+			var functions = mapping.GetAttributes<Sql.ExtensionAttribute>(methodCall.Method.ReflectedTypeEx(),
+				methodCall.Method,
+				f => f.Configuration);
+			return functions.Any();
+		}
+
 		public static bool IsQueryable(this MethodCallExpression method, string name)
 		{
 			return method.Method.Name == name && method.IsQueryable();
@@ -1034,7 +1050,7 @@ namespace LinqToDB.Expressions
 						var call = (MethodCallExpression)expression;
 						var expr = call.Object;
 
-						if (expr == null && (call.IsQueryable() || call.IsAggregate(mapping) || call.IsAssociation(mapping) || call.Method.IsSqlPropertyMethodEx()) && call.Arguments.Count > 0)
+						if (expr == null && (call.IsQueryable() || call.IsAggregate(mapping) || call.IsExtensionMethod(mapping) || call.IsAssociation(mapping) || call.Method.IsSqlPropertyMethodEx()) && call.Arguments.Count > 0)
 							expr = call.Arguments[0];
 
 						if (expr != null)
