+++ Source/LinqToDB/Linq/Builder/BuildInfo.cs
@@ -36,7 +36,6 @@ namespace LinqToDB.Linq.Builder
 		public bool                 CreateSubQuery { get; set; }
 		public JoinType             JoinType       { get; set; }
 		public bool                 IsSubQuery => Parent != null;
-		public bool                 IsChain        { get; set; }
 
 		private bool _isAssociationBuilt;
 		public  bool  IsAssociationBuilt
