+++ Source/LinqToDB/Linq/Builder/MethodChainBuilder.cs
@@ -16,37 +16,23 @@ namespace LinqToDB.Linq.Builder
 		protected override bool CanBuildMethodCall(ExpressionBuilder builder, MethodCallExpression methodCall, BuildInfo buildInfo)
 		{
 			var functions = Sql.ExtensionAttribute.GetExtensionAttributes(methodCall, builder.MappingSchema);
-			return functions.Any(f => f.ChainPrecedence >= 0);
+			return functions.Any();
 		}
 
 		protected override IBuildContext BuildMethodCall(ExpressionBuilder builder, MethodCallExpression methodCall, BuildInfo buildInfo)
 		{
 			var functions = Sql.ExtensionAttribute.GetExtensionAttributes(methodCall, builder.MappingSchema);
 
-			var chain = Sql.ExtensionAttribute.BuildFunctionsChain(builder.MappingSchema, methodCall);
-			IBuildContext sequence = null;
+			var root = methodCall.SkipMethodChain(builder.MappingSchema);
 
-			foreach (var expression in chain)
+			// evaluating IQueryableContainer
+			while (root.NodeType == ExpressionType.Constant && typeof(Sql.IQueryableContainer).IsSameOrParentOf(root.Type))
 			{
-				if (expression is MethodCallExpression mc)
-				{
-					if (mc.Arguments.Count > 0)
-					{
-						if (typeof(IEnumerable<>).IsSameOrParentOf(mc.Arguments[0].Type))
-							sequence = builder.BuildSequence(new BuildInfo(buildInfo, mc.Arguments[0]) {IsChain = true});
-						else if (typeof(Sql.IQueryableContainer).IsSameOrParentOf(mc.Arguments[0].Type))
-							sequence = builder.BuildSequence(new BuildInfo(buildInfo,
-								((Sql.IQueryableContainer)mc.Arguments[0].EvaluateExpression()).Query.Expression) {IsChain = true});
-					}
-				}
+				root = ((Sql.IQueryableContainer)root.EvaluateExpression()).Query.Expression;
+				root = root.SkipMethodChain(builder.MappingSchema);
 			}
 
-			if (sequence == null)
-				throw new LinqToDBException($"{methodCall} can not be converted to SQL.");
-
-			if (buildInfo.IsChain)
-				// it means that function is just sequence provider
-				return sequence;
+			var sequence = builder.BuildSequence(new BuildInfo(buildInfo, root) { CreateSubQuery = true });
 
 			var finalFunction = functions.First();
